% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_genes}
\alias{BRCA_genes}
\title{BRCA genes}
\format{
BRCA_genes: A matrix object with 4819 BRCA related genes (including lncRNAs and mRNAs).
}
\description{
BRCA genes
}
\details{
The BRCA related lncRNAs are from LncRNADisease v2.0, Lnc2Cancer v2.0
and MNDR v2.0. The BRCA related mRNAs are from DisGeNET v5.0 and COSMIC v86.
}
\references{
Bao Z, Yang Z, Huang Z, Zhou Y, Cui Q, Dong D. LncRNADisease 2.0: an updated database of long non-coding RNA-associated diseases. Nucleic Acids Res. 2019;47(D1):D1034-D1037.

Cui T, Zhang L, Huang Y, Yi Y, Tan P, Zhao Y, Hu Y, Xu L, Li E, Wang D. MNDR v2.0: an updated resource of ncRNA-disease associations in mammals. Nucleic Acids Res. 2018;46(D1):D371-D374.

Gao Y, Wang P, Wang Y, Ma X, Zhi H, Zhou D, Li X, Fang Y, Shen W, Xu Y, Shang S, Wang L, Wang L, Ning S, Li X. Lnc2Cancer v2.0: updated database of experimentally supported
long non-coding RNAs in human cancers. Nucleic Acids Res. 2019;47(D1):D1028-D1033.

Forbes SA, Beare D, Boutselakis H, Bamford S, Bindal N, Tate J, Cole CG, Ward S, Dawson E, Ponting L, Stefancsik R, Harsha B, Kok CY, Jia M, Jubb H, Sondka Z, Thompson S, De T,
Campbell PJ. COSMIC: somatic cancer genetics at high-resolution. Nucleic Acids Res. 2017;45(D1):D777-D783.

Piñero J, Bravo À, Queralt-Rosinach N, Gutiérrez-Sacristán A, Deu-Pons J, Centeno E, García-García J, Sanz F, Furlong LI. DisGeNET: a comprehensive platform integrating information
on human disease-associated genes and variants. Nucleic Acids Res. 2017;45(D1):D833-D839.

Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. miRSM: an R package
to infer and analyse miRNA sponge modules in heterogeneous data. RNA Biol. 2021;18(12):2308-2320.
}
