% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_tar_cancer}
\alias{bulk_tar_cancer}
\title{Bulk target gene expression data in breast cancer}
\format{
bulk_tar_cancer: A data frame object with 6 BRCA samples (columns) and 22427 target genes (rows).
}
\description{
Bulk target gene expression data in breast cancer
}
\details{
The target gene expression data in breast cancer is obtained from TCGA (http://cancergenome.nih.gov/).
}
