% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{CSN}
\alias{CSN}
\title{CSN}
\usage{
CSN(
  reg_cancer,
  tar_cancer,
  boxsize = 0.1,
  bootstrap = FALSE,
  bootstrap_betw_point = 5,
  bootstrap_num = 100,
  p.value.cutoff = 0.05,
  num.cores = 2
)
}
\arguments{
\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{boxsize}{Size of neighborhood (0.1 in default).}

\item{bootstrap}{A logic value, TRUE for bootstrap. If the number of samples is less than 100, boostrap is recommended.}

\item{bootstrap_betw_point}{The number of interpolation points between each cell (5 in default), bootstrap_betw_point = 0 is used to compute.}

\item{bootstrap_num}{The number of bootstrapping for interpolating pseudo-cells.}

\item{p.value.cutoff}{Significance p-value for identifying sample-specific regulatory network.}

\item{num.cores}{The number of cores to run in parallel.}
}
\value{
A list object containing regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the CSN (cell-specific network) method, adapted from the CSmiR method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_CSN_res <- CSN(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], boxsize = 0.1, p.value.cutoff = 0.05)

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_CSN_res <- CSN(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], boxsize = 0.1, p.value.cutoff = 0.05)

# ST data
data(ST_ncR_tar_cancer)
ST_CSN_res <- CSN(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5], boxsize = 0.1, p.value.cutoff = 0.05)

}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.2021;22(1):578.
Dai H, Li L, Zeng T, Chen L. Cell-specific network constructed by single-cell RNA sequencing data. Nucleic Acids Res. 2019;47(11):e62.
}
