% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_ncRNAs}
\alias{hub_ncRNAs}
\title{hub_ncRNAs}
\format{
hub_ncRNAs: A list object with hub ncRNAs existed in five ncRNA-target regulatory networks predicted by the SCS method
}
\description{
hub_ncRNAs
}
\details{
A list of hub ncRNAs existed in five ncRNA-target regulatory networks as a demo
}
