% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Hub_sim_hc}
\alias{Hub_sim_hc}
\title{Hub_sim_hc}
\usage{
Hub_sim_hc(hub, method = "Simpson")
}
\arguments{
\item{hub}{List object, a list of sample-specific hubs.}

\item{method}{Methods for calculating similatiry between two two lists of sample-specific hubs, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Hierarchical cluster analysis result of samples.
}
\description{
Performing hierarchical clustering analysis of samples using the calculated sample-sample similarity matrix in terms of sample-specific hubs
}
\examples{
data(hub_ncRNAs)
res <- Hub_sim_hc(hub_ncRNAs)
}
