% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{LIONESS}
\alias{LIONESS}
\title{LIONESS}
\usage{
LIONESS(reg_cancer, tar_cancer, cormethod = "pearson", num.cores = 2)
}
\arguments{
\item{reg_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{cormethod}{A character string indicating which type of correlation coefficient is to be computed.
One of 'pearson' (default), 'kendall', or 'spearman' can be used.}

\item{num.cores}{Integer, number of cores to use for parallel processing. Default is 2.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the LIONESS (Linear Interpolation to Obtain Network Estimates for Single Samples) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_LIONESS_res <- LIONESS(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5],
                            cormethod = 'pearson')

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_LIONESS_res <- LIONESS(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5],
                             cormethod = 'pearson')

# ST data
data(ST_ncR_tar_cancer)
ST_LIONESS_res <- LIONESS(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5],
                          cormethod = 'pearson')

}
\references{
Kuijjer ML, Tung MG, Yuan G, Quackenbush J, Glass K. Estimating Sample-Specific Regulatory Networks. iScience. 2019;14:226-240.
}
