% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{miRTarget_Groundtruth}
\alias{miRTarget_Groundtruth}
\title{Groundtruth of miRNA-target interactions}
\format{
miRTarget_Groundtruth: A data frame object with 2,956,530 miRNA-target interactions.
}
\description{
Groundtruth of miRNA-target interactions
}
\details{
The groundtruth of miRNA-target interactions are from miRTarBase v10.0 and TarBase v9.0.
}
\references{
Cui S, Yu S, Huang HY, Lin YC, Huang Y, Zhang B, Xiao J, Zuo H, Wang J, Li Z, Li G, Ma J, Chen B, Zhang H, Fu J, Wang L,
Huang HD. miRTarBase 2025: updates to the collection of experimentally validated microRNA-target interactions. Nucleic Acids Res. 2025;53(D1):D147-D156.

Skoufos G, Kakoulidis P, Tastsoglou S, Zacharopoulou E, Kotsira V, Miliotis M, Mavromati G, Grigoriadis D, Zioga M, Velli A, Koutou I,
Karagkouni D, Stavropoulos S, Kardaras FS, Lifousi A, Vavalou E, Ovsepian A, Skoulakis A, Tasoulis SK, Georgakopoulos SV, Plagianakos VP, Hatzigeorgiou AG.
TarBase-v9.0 extends experimentally supported miRNA-gene interactions to cell-types and virally encoded miRNAs. Nucleic Acids Res. 2024;52(D1):D304-D310.
}
