% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Module_sim_hc}
\alias{Module_sim_hc}
\title{Module_sim_hc}
\usage{
Module_sim_hc(module, method = "Simpson")
}
\arguments{
\item{module}{List object, a list of sample-specific module groups. Each module group contain a list of modules.}

\item{method}{Methods for calculating similatiry between two lists of sample-specific modules, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Hierarchical cluster analysis result of samples.
}
\description{
Performing hierarchical clustering analysis of samples using the calculated sample-sample similarity matrix in terms of sample-specific modules
}
