% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_ncR}
\alias{mut_ncR}
\title{mutation data}
\format{
mut_ncR: A character object with 158 SNV genes.
}
\description{
mutation data
}
\details{
The mutation genes are collected from miRNASNP v3.0 (http://bioinfo.life.hust.edu.cn/miRNASNP/#!/) and lncRNASNP v3.0 (http://bioinfo.life.hust.edu.cn/lncRNASNP#!/) databases.
}
\references{
Liu CJ, Fu X, Xia M, Zhang Q, Gu Z, Guo AY. miRNASNP-v3: a comprehensive database for SNPs and disease-related variations in miRNAs and miRNA targets. Nucleic Acids Res. 2021;49(D1):D1276-D1281.

Yang Y, Wang D, Miao YR, Wu X, Luo H, Cao W, Yang W, Yang J, Guo AY, Gong J. lncRNASNP v3: an updated database for functional variants in long non-coding RNAs. Nucleic Acids Res. 2023;51(D1):D192-D198.
}
