% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{network}
\format{
network: A list object with five ncRNA-target regulatory networks predicted by the SCS method
}
\description{
network
}
\details{
A list of five ncRNA-target regulatory networks as a demo
}
