% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_hub}
\alias{Network_hub}
\title{Network_hub}
\usage{
Network_hub(network, directed = TRUE, top_percentage = 0.2)
}
\arguments{
\item{network}{List object: a list of of sample-specific networks.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{top_percentage}{A number indicating the top percenatge of hub regulators, default 0.2.}
}
\value{
List object: a list of hub regulators for each sample.
}
\description{
Identifying sample-specific hub regulators
}
\examples{
data(network)
hubs <- Network_hub(network)
}
\references{
Zhang J, Liu L, Xu T, Zhang W, Zhao C, Li S, Li J, Rao N, Le TD. Exploring cell-specific miRNA regulation with single-cell miRNA-mRNA co-sequencing data. BMC Bioinformatics.2021; 22(1):578.
}
