% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_module}
\alias{Network_module}
\title{Network_module}
\usage{
Network_module(network_graph, method = "MCL", directed = FALSE, modulesize = 3)
}
\arguments{
\item{network_graph}{An igraph object, a list of sample-specific networks.}

\item{method}{Possible methods include FN, MCL, LINKCOMM, MCODE, betweenness, infomap, prop, eigen, louvain, walktrap.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{modulesize}{The size of modules, e.g. 3.}
}
\value{
A list object: a list of sample-specific modules.
}
\description{
Predicting sample-specific modules
}
\examples{
data(bulk_ncR_tar_normal_cancer)
bulk_LIONESS_res <- LIONESS(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], cormethod = 'pearson')
Network_module_res <- Network_module(bulk_LIONESS_res)

}
