% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_sim_hc}
\alias{Network_sim_hc}
\title{Network_sim_hc}
\usage{
Network_sim_hc(network, directed = TRUE, method = "Simpson")
}
\arguments{
\item{network}{List object, a list of sample-specific networks.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{method}{Methods for calculating similatiry between two two lists of sample-specific networks, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Hierarchical cluster analysis result of samples.
}
\description{
Performing hierarchical clustering analysis of samples using the calculated sample-sample similarity matrix in terms of sample-specific networks
}
\examples{
data(network)
res <- Network_sim_hc(network)
}
