% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Scan.interp}
\alias{Scan.interp}
\title{Scan.interp}
\usage{
Scan.interp(
  reg_cancer,
  tar_cancer,
  method = c("pearson", "spearman", "kendall"),
  p.value.cutoff = 0.05,
  num.cores = 2
)
}
\arguments{
\item{reg_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{method}{Methods for calculating correlations, one of 'pearson', 'spearman', 'kendall'.}

\item{p.value.cutoff}{Significance level for the identified regulator-target interactions.}

\item{num.cores}{Number of CPU cores when parallel computation.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Identifying sample-specific ncRNA regulation using Scan (Sample-speCific miRNA regulAtioN) and linear interpolation strategy
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_Scan.interp_res <- Scan.interp(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], method = 'pearson', p.value.cutoff = 0.05)

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_Scan.interp_res <- Scan.interp(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], method = 'pearson', p.value.cutoff = 0.05)

# ST data
data(ST_ncR_tar_cancer)
ST_Scan.interp_res <- Scan.interp(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5], method = 'pearson', p.value.cutoff = 0.05)

}
\references{
Zhang J, Liu L, Wei X, Zhao C, Luo Y, Li J, Le TD. Scanning sample-specific miRNA regulation from bulk and single-cell RNA-sequencing data. BMC Biol. 2024;22(1):218.
}
