% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Scan.perturb}
\alias{Scan.perturb}
\title{Scan.perturb}
\usage{
Scan.perturb(
  reg_cancer,
  tar_cancer,
  method = c("Pearson", "Spearman", "Kendall"),
  p.value.cutoff = 0.05,
  num.cores = 2
)
}
\arguments{
\item{reg_cancer}{miRNA expression data with columns are samples and rows are regulators.}

\item{tar_cancer}{mRNA expression data with columns are samples and rows are mRNAs.}

\item{method}{Methods for calculating correlations, one of "Pearson", "Spearman", "Kendall".}

\item{p.value.cutoff}{Significance level for the identified regulator-target interactions.}

\item{num.cores}{Number of CPU cores when parallel computation.}
}
\value{
A list object containing RNA-RNA interactions for each sample.
}
\description{
Identifying sample-specific miRNA regulation using Scan (Sample-speCific miRNA regulAtioN) and statistical perturbation strategy
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_Scan.perturb_res <- Scan.perturb(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], method = "Pearson", p.value.cutoff = 0.05)

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_Scan.perturb_res <- Scan.perturb(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], method = "Pearson", p.value.cutoff = 0.05)

# ST data
data(ST_ncR_tar_cancer)
ST_Scan.perturb_res <- Scan.perturb(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5], method = "Pearson", p.value.cutoff = 0.05)

}
\references{
Zhang J, Liu L, Wei X, Zhao C, Luo Y, Li J, Le TD. Scanning sample-specific miRNA regulation from bulk and single-cell RNA-sequencing data. BMC Biol. 2024;22(1):218.
}
