% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA_ncR_normal}
\alias{scRNA_ncR_normal}
\title{single-cell ncRNA expression data in normal breast}
\format{
scRNA_ncR_normal: A data frame object with 2430 barcodes of normal breast cells (columns) and 343 ncRNAs (rows).
}
\description{
single-cell ncRNA expression data in normal breast
}
\details{
The matched miRNA and lncRNA expression data in normal breast cells is obtained from GEO (GSE153889).
The data is pre-processed by using Seurat R package.
The matched miRNA and lncRNA expression data is regarded as single-cell ncRNA expression data in normal breast.
}
\references{
Martin Carli JF, Trahan GD, Jones KL, Hirsch N, Rolloff KP, Dunn EZ, Friedman JE, Barbour LA, Hernandez TL, MacLean PS, Monks J,
McManaman JL, Rudolph MC. Single Cell RNA Sequencing of Human Milk-Derived Cells Reveals Sub-Populations of Mammary Epithelial Cells with Molecular
Signatures of Progenitor and Mature States: a Novel, Non-invasive Framework for Investigating Human Lactation Physiology. J Mammary Gland Biol Neoplasia.
2020;25(4):367-387.
}
