% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA_tar_count_cancer}
\alias{scRNA_tar_count_cancer}
\title{Single-cell target gene count data in breast cancer}
\format{
scRNA_tar_count_cancer: A data frame object with 736 barcodes of CID3946 breast cancer cells (columns) and 9323 target genes (rows).
}
\description{
Single-cell target gene count data in breast cancer
}
\details{
The target gene count data is obtained from Wu et al.
The target gene expression data is regarded as single-cell target gene count data in breast cancer.
}
\references{
Wu SZ, Al-Eryani G, Roden DL, et al.  A single-cell and spatially resolved atlas of human breast cancers. Nat Genet, 2021;53(9):1334-1347.
}
