% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Sim.hub}
\alias{Sim.hub}
\title{Sim.hub}
\usage{
Sim.hub(hub1, hub2, method = "Simpson")
}
\arguments{
\item{hub1}{The first list of sample-specific hubs.}

\item{hub2}{The second list of sample-specific hubs.}

\item{method}{Methods for calculating similatiry between two two lists of sample-specific hubs, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Matrix object, a sample-sample similarity matrix in terms of sample-specific hubs.
}
\description{
Calculating sample-sample similarity matrix between two lists of sample-specific hubs
}
\examples{
data(hub_ncRNAs)
res <- Sim.hub(hub_ncRNAs, hub_ncRNAs)
}
