% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Sim.module}
\alias{Sim.module}
\title{Sim.module}
\usage{
Sim.module(Module1, Module2, method = "Simpson")
}
\arguments{
\item{Module1}{List object, the first list of modules.}

\item{Module2}{List object, the second list of modules.}

\item{method}{Methods for calculating similatiry between two modules, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Matrix object, a sample-sample similarity matrix in terms of sample-specific modules.
}
\description{
Calculating sample-sample similarity matrix between two lists of sample-specific modules
}
