% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{SINUM}
\alias{SINUM}
\title{SINUM}
\usage{
SINUM(reg_cancer, tar_cancer, boxsize = 0.2, zscore.cutoff = 0, num.cores = 2)
}
\arguments{
\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{boxsize}{Size of neighborhood (0.2 in default).}

\item{zscore.cutoff}{Zscore cutoff for identifying sample-specific regulatory network.}

\item{num.cores}{The number of cores to run in parallel.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the SINUM (SIngle-cell Network Using Mutual information) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_SINUM_res <- SINUM(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], boxsize = 0.2, zscore.cutoff = 0)

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_SINUM_res <- SINUM(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], boxsize = 0.2, zscore.cutoff = 0)

# ST data
data(ST_ncR_tar_cancer)
ST_SINUM_res <- SINUM(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5], boxsize = 0.2, zscore.cutoff = 0)

}
\references{
Chang LY, Hao TY, Wang WJ, Lin CY. Inference of single-cell network using mutual information for scRNA-seq data analysis. BMC Bioinformatics. 2024;25(Suppl 2):292.
}
