% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{SWEET}
\alias{SWEET}
\title{SWEET}
\usage{
SWEET(reg_cancer, tar_cancer, dismethod = "euclidean", threshold = 0.7)
}
\arguments{
\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{dismethod}{A character string indicating which type of distance is to be computed. One of 'euclidean' (default), or 'correlation' can be used.}

\item{threshold}{The cutoff of correlation strength.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the SWEET (Sample-specific weighted correlation network) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_SWEET_res <- SWEET(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], dismethod = 'euclidean')

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_SWEET_res <- SWEET(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], dismethod = 'euclidean')

# ST data
data(ST_ncR_tar_cancer)
ST_SWEET_res <- SWEET(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5], dismethod = 'euclidean')

}
\references{
Chen HH, Hsueh CW, Lee CH, Hao TY, Tu TY, Chang LY, Lee JC, Lin CY. SWEET: a single-sample network inference method for deciphering individual features in disease. Brief Bioinform. 2023, 24(2):bbad032.
}
