% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFTarget_Groundtruth}
\alias{TFTarget_Groundtruth}
\title{Groundtruth of TF-target interactions}
\format{
TFTarget_Groundtruth: A data frame object with 43,178 TF-target interactions.
}
\description{
Groundtruth of TF-target interactions
}
\details{
The groundtruth of TF-target interactions are from CollecTRI.
}
\references{
Müller-Dott S, Tsirvouli E, Vazquez M, Ramirez Flores RO, Badia-I-Mompel P, Fallegger R, Türei D, Lægreid A, Saez-Rodriguez J.
Expanding the coverage of regulons from high-confidence prior knowledge for accurate estimation of transcription factor activities. Nucleic Acids Res. 2023 Nov 10;51(20):10934-10949.
}
