% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{BONOBO}
\alias{BONOBO}
\title{BONOBO}
\usage{
BONOBO(reg_cancer, tar_cancer, p.value.cutoff = 0.05)
}
\arguments{
\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{p.value.cutoff}{pvalue cutoff for identifying sample-specific regulatory network.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the BONOBO (Bayesian Optimized Networks Obtained By assimilating Omics data) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_BONOBO_res <- BONOBO(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5])

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_BONOBO_res <- BONOBO(scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5])

# ST data
data(ST_ncR_tar_cancer)
ST_BONOBO_res <- BONOBO(ST_ncR_cancer[1:100, 1:5], ST_tar_cancer[1:500, 1:5])

}
\references{
Saha E, Fanfani V, Mandros P, Ben Guebila M, Fischer J, Shutta KH, DeMeo DL, Lopes-Ramos CM, Quackenbush J. Bayesian inference of sample-specific coexpression networks. Genome Res. 2024;34(9):1397-1410.
}
