% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_ncR_normal}
\alias{bulk_ncR_normal}
\title{Bulk ncRNA expression data in normal breast}
\format{
bulk_ncR_normal: A data frame object with six normal samples (columns) and 4929 ncRNAs (rows).
}
\description{
Bulk ncRNA expression data in normal breast
}
\details{
The matched miRNA and lncRNA expression data in normal breast is obtained from TCGA (http://cancergenome.nih.gov/).
}
