% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulk_tar_normal}
\alias{bulk_tar_normal}
\title{Bulk target gene expression data in normal breast}
\format{
bulk_tar_normal: A dataframe object with six normal samples (columns) and 22427 target genes (rows).
}
\description{
Bulk target gene expression data in normal breast
}
\details{
The matched lncRNA and mRNA expression data in normal breast is obtained from TCGA (http://cancergenome.nih.gov/).
}
