% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_ncR}
\alias{mut_ncR}
\title{Mutation data}
\format{
mut_ncR: A character object with 158 mutation lncRNAs.
}
\description{
Mutation data
}
\details{
The mutation lncRNAs are collected from lncRNASNP v3.0 (http://gong_lab.hzau.edu.cn/lncRNASNP3/).
}
\references{
Yang Y, Wang D, Miao YR, Wu X, Luo H, Cao W, Yang W, Yang J, Guo AY, Gong J. lncRNASNP v3: an updated database for functional variants in long non-coding RNAs. Nucleic Acids Res. 2023;51(D1):D192-D198.
}
