% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_module}
\alias{Network_module}
\title{Network_module}
\usage{
Network_module(network, method = "MCL", directed = TRUE, modulesize = 3)
}
\arguments{
\item{network}{List object, a list of sample-specific networks.}

\item{method}{Possible methods include FN, MCL, LINKCOMM, MCODE, betweenness, infomap, prop, eigen, louvain, walktrap.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{modulesize}{The size of modules, e.g. 3.}
}
\value{
List object: a list of sample-specific modules.
}
\description{
Predicting sample-specific modules
}
\examples{
data(network)
Network_module_res <- Network_module(network)
}
