% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_sim_hc}
\alias{Network_sim_hc}
\title{Network_sim_hc}
\usage{
Network_sim_hc(network_graph, directed = FALSE, method = "Simpson")
}
\arguments{
\item{network_graph}{An igraph object, a list of sample-specific networks.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{method}{Methods for calculating similatiry between two two lists of sample-specific networks, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Hierarchical cluster analysis result of samples.
}
\description{
Performing hierarchical clustering analysis of samples using the calculated sample-sample similarity matrix in terms of sample-specific networks
}
\examples{
data(bulk_ncR_tar_normal_cancer)
bulk_LIONESS_res <- LIONESS(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], cormethod = 'pearson')
res <- Network_sim_hc(bulk_LIONESS_res)
plot(res)

}
