% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Network_validation}
\alias{Network_validation}
\title{Network_validation}
\usage{
Network_validation(network, groundtruth)
}
\arguments{
\item{network}{List object: a list of sample-specific networks.}

\item{groundtruth}{A data frame object, experimentally validated RNA-RNA interactions}
}
\value{
List object: a list of validated RNA-RNA interactions for each sample.
}
\description{
Validating sample-specific network
}
