% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{PairedSSN}
\alias{PairedSSN}
\title{PairedSSN}
\usage{
PairedSSN(
  reg_normal,
  tar_normal,
  reg_cancer,
  tar_cancer,
  cormethod = "pearson",
  p.value.cutoff = 0.05
)
}
\arguments{
\item{reg_normal}{A data frame object, the input regulator expression data in normal samples, columns are samples and rows are regulators.}

\item{tar_normal}{A data frame object, the input target gene expression data in normal samples, columns are samples and rows are target genes.}

\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{cormethod}{A character string indicating which type of correlation coefficient is to be computed. One of 'pearson' (default), 'kendall', or 'spearman' can be used.}

\item{p.value.cutoff}{Significance p-value for identifying sample-specific regulatory network}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the Paired-SSN (paired single sample network) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_PaiedSSN_res <- PairedSSN(bulk_ncR_normal[1:100, 1:5], bulk_tar_normal[1:500, 1:5], bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], cormethod = 'pearson')

# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_PaiedSSN_res <- PairedSSN(scRNA_ncR_normal[1:100, 1:5], scRNA_tar_normal[1:500, 1:5], scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5], cormethod = 'pearson')

}
\references{
Guo WF, Zhang SW, Zeng T, Li Y, Gao J, Chen L. A novel network control model for identifying personalized driver genes in cancer. PLoS Comput Biol. 2019;15(11):e1007520.
}
