% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA_ncR_cancer}
\alias{scRNA_ncR_cancer}
\title{single-cell ncRNA expression data in breast cancer}
\format{
scRNA_ncR_cancer: A data frame object with 736 barcodes of CID3946 breast cancer cells (columns) and 147 ncRNAs (rows).
}
\description{
single-cell ncRNA expression data in breast cancer
}
\details{
The matched miRNA and lncRNA expression data in breast cancer cells is obtained from Wu et al.
The data is pre-processed by using Seurat R package.
The matched miRNA and lncRNA expression data is regarded as single-cell ncRNA expression data in breast cancer.
}
\references{
Wu SZ, Roden DL, Junankar S et al. A single-cell and spatially resolved atlas of human breast cancers. Nat Genet, 2021;53(9):1334-1347.
}
