% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRNA_tar_normal}
\alias{scRNA_tar_normal}
\title{Single-cell target gene expression data in normal breast}
\format{
scRNA_tar_normal: A data frame object with 2430 barcodes of normal breast cells (columns) and 11226 target genes (rows).
}
\description{
Single-cell target gene expression data in normal breast
}
\details{
The target gene expression data in normal breast cells is obtained from GEO (GSE15388).
The target gene expression data is regarded as single-cell target gene expression data in normal breast.
}
\references{
Martin Carli JF, Trahan GD, Jones KL, Hirsch N, Rolloff KP, Dunn EZ, Friedman JE, Barbour LA, Hernandez TL, MacLean PS, Monks J, McManaman JL,
Rudolph MC. Single Cell RNA Sequencing of Human Milk-Derived Cells Reveals Sub-Populations of Mammary Epithelial Cells with Molecular Signatures of Progenitor
and Mature States: a Novel, Non-invasive Framework for Investigating Human Lactation Physiology. J Mammary Gland Biol Neoplasia. 2020;25(4):367-387.
}
