% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{SCS}
\alias{SCS}
\title{SCS}
\usage{
SCS(mut_reg, reg_cancer, tar_cancer, cormethod = "pearson", number = 50)
}
\arguments{
\item{mut_reg}{A matrix or data frame object, containing a list of mutation regulators.}

\item{reg_cancer}{A data frame object, the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A data frame object, the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{cormethod}{A character string indicating which type of correlation coefficient is to be computed. One of 'pearson' (default), 'kendall', or 'spearman' can be used.}

\item{number}{The number of RWR with random matrix.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with SCS (single-sample controller strategy) method
}
\examples{
# scRNA data
data(scRNA_ncR_tar_normal_cancer)
data(mut_ncR)
scRNA_SCS_res <- SCS(mut_ncR, scRNA_ncR_cancer[, 1:5], scRNA_tar_cancer[1:500, 1:5], cormethod = 'pearson', number = 10)

# ST data
data(ST_ncR_tar_cancer)
data(mut_ncR)
ST_SCS_res <- SCS(mut_ncR, ST_ncR_cancer[, 1:5], ST_tar_cancer[1:500, 1:5], cormethod = 'pearson', number = 10)

}
\references{
Guo WF, Zhang SW, Liu LL, Liu F, Shi QQ, Zhang L, Tang Y, Zeng T, Chen L. Discovering personalized driver mutation profiles of single samples in cancer by network control strategy. Bioinformatics. 2018;34(11):1893-1903.
}
