% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Sim.hub}
\alias{Sim.hub}
\title{Sim.hub}
\usage{
Sim.hub(hub1, hub2, method = "Simpson")
}
\arguments{
\item{hub1}{The first list of sample-specific hubs.}

\item{hub2}{The second list of sample-specific hubs.}

\item{method}{Methods for calculating similarity between two two lists of sample-specific hubs, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
A matrix object, a sample-sample similarity matrix in terms of sample-specific hubs.
}
\description{
Calculating sample-sample similarity matrix between two lists of sample-specific hubs
}
\examples{
data(bulk_ncR_tar_normal_cancer)
bulk_LIONESS_res <- LIONESS(bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5], cormethod = 'pearson')
hubs <- Network_hub(bulk_LIONESS_res)
res <- Sim.hub(hubs, hubs)

}
