% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{Sim.network}
\alias{Sim.network}
\title{Sim.network}
\usage{
Sim.network(net1, net2, directed = TRUE, method = "Simpson")
}
\arguments{
\item{net1}{The first list of sample-specific networks.}

\item{net2}{The second list of sample-specific networks.}

\item{directed}{A logical value indicating a network directed (TRUE) or undirected (FALSE).}

\item{method}{Methods for calculating similatiry between two two lists of sample-specific networks, select one of three mrthods (Simpson, Jaccard and Lin). Default method is Simpson.}
}
\value{
Matrix object, a sample-sample similarity matrix in terms of sample-specific networks.
}
\description{
Calculating sample-sample similarity matrix between two lists of sample-specific networks
}
\examples{
data(network)
res <- Sim.network(network, network)
}
