% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S2N.R
\name{SSN}
\alias{SSN}
\title{SSN}
\usage{
SSN(
  reg_normal,
  tar_normal,
  reg_cancer,
  tar_cancer,
  cormethod = "pearson",
  p.value.cutoff = 0.05,
  num.cores = 2
)
}
\arguments{
\item{reg_normal}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input regulator expression data in normal samples, columns are samples and rows are regulators.}

\item{tar_normal}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input target gene expression data in normal samples, columns are samples and rows are target genes.}

\item{reg_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input regulator expression data in tumor samples, columns are samples and rows are regulators.}

\item{tar_cancer}{A dataframe or matrix or SummarizedExperiment or SingleCellExperiment or SpatialExperiment object,
the input target gene expression data in tumor samples, columns are samples and rows are target genes.}

\item{cormethod}{A character string indicating which type of correlation coefficient is to be computed.
One of 'pearson' (default), 'kendall', or 'spearman' can be used.}

\item{p.value.cutoff}{Significance p-value for identifying sample-specific regulatory network.}

\item{num.cores}{Integer, number of cores to use for parallel processing. Default is 2.}
}
\value{
An igraph object: a list of regulator-target interactions for each sample.
}
\description{
Sample-specific network inferred with the SSN (single-sample network) method
}
\examples{
# bulk data
data(bulk_ncR_tar_normal_cancer)
bulk_SSN_res <- SSN(bulk_ncR_normal[1:100, 1:5], bulk_tar_normal[1:500, 1:5],
                    bulk_ncR_cancer[1:100, 1:5], bulk_tar_cancer[1:500, 1:5],
                    cormethod = 'pearson')
# scRNA data
data(scRNA_ncR_tar_normal_cancer)
scRNA_SSN_res <- SSN(scRNA_ncR_normal[1:100, 1:5], scRNA_tar_normal[1:500, 1:5],
                     scRNA_ncR_cancer[1:100, 1:5], scRNA_tar_cancer[1:500, 1:5],
                     cormethod = 'pearson')

}
\references{
Liu X, Wang Y, Ji H, Aihara K, Chen L. Personalized characterization of diseases using sample-specific networks. Nucleic Acids Res. 2016;44(22):e164.
}
