% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_ncR_cancer}
\alias{ST_ncR_cancer}
\title{Spatial transcriptomics data of lncRNA in breast cancer}
\format{
ST_ncR_cancer: A data frame object with 1057 barcodes of CID44971 breast cancer samples (columns) and 2109 lncRNAs (rows).
}
\description{
Spatial transcriptomics data of lncRNA in breast cancer
}
\details{
The spatial transcriptomics data of lncRNAs in breast cancer is obtained from Wu et al.
The spatial transcriptomics data is pre-processed by using Seurat R package.
The lncRNA expression data is regarded as spatial transcriptomics data of lncRNA in breast cancer.
}
\references{
Wu SZ, Roden DL, Junankar S et al. A single-cell and spatially resolved atlas of human breast cancers. Nat Genet, 2021;53(9):1334-1347.
}
