% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_tar_cancer}
\alias{ST_tar_cancer}
\title{Spatial transcriptomics data of mRNA in breast cancer}
\format{
ST_tar_cancer: A data frame object with 1057 barcodes of CID44971 breast cancer samples (columns) and 14801 mRNAs (rows).
}
\description{
Spatial transcriptomics data of mRNA in breast cancer
}
\details{
The spatial transcriptomics data of mRNAs in breast cancer is obtained from Wu et al.
The spatial transcriptomics data is pre-processed by using Seurat R package.
The mRNA expression data is regarded as spatial transcriptomics data of mRNA in breast cancer.
}
\references{
Wu SZ, Roden DL, Junankar S et al. A single-cell and spatially resolved atlas of human breast cancers. Nat Genet, 2021;53(9):1334-1347.
}
